Hide description when no descriptions are provided
